% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Intermediate_functions.R
\name{cam.order}
\alias{cam.order}
\title{Compute the camera order}
\usage{
cam.order(time_df)
}
\arguments{
\item{time_df}{a numerical dataframe containing the minimal time 
needed for an individual of the studied species to go from a camera to 
another camera.There are as many rows as there are cameras and there are
as many columns as there are cameras, thus the dataframe is symmetrical
and the diagonal is filled with 0. This dataframe is the output of the 
\code{compute.cam.time} function.}
}
\value{
a dataframe with one row containing the camera order to follow
}
\description{
This function rated the cameras in the following order: the first camera
is the central one (lowest mean distance across all other camera) (if several
camera possible, then classify in the order they come in the abund_df, it is 
not that important as we test all possible paths). Then other cameras are 
rated according to their minimal distances to the camers already chosen.
}
\examples{
time_df <- data.frame("A" = c(0,1,3,3), "B" = c(1,0,2,2), 
                      "C" = c(3,3,0,2), "D" = c(3,2,2,0))
rownames(time_df) <- c("A", "B", "C", "D")

order_cam <- cam.order(time_df)

}
