% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_SmaxN_big_interval.R
\name{compute.cam.order}
\alias{compute.cam.order}
\title{Compute the order in which cameras must be looked at in a bloc (smallest
to highest distances)}
\usage{
compute.cam.order(time_df)
}
\arguments{
\item{time_df}{a numerical dataframe containing the minimal time 
needed for an individual of the studied species to go from a camera to 
another camera.There are as many rows as there are cameras and there are
as many columns as there are cameras, thus the dataframe is symmetrical
and the diagonal is filled with 0. This dataframe is the output of the 
\code{compute.cam.time} function.}
}
\value{

}
\description{
This function computes a dataframe which contains one or several rows which...
camera names in the order they must be looked at. If several camera ...
pairs have the same distance, then their will be several rows... 
If all camera pairs have different distances, then there will be one row ... 
with only one camera order to test
}
