% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_SmaxN_big_interval.R
\name{compute.SmaxN.bigUI}
\alias{compute.SmaxN.bigUI}
\title{Compute the SmaxN for a timespan ie here the big interval of a given timestep}
\usage{
compute.SmaxN.bigUI(abund_df, value, timestep, time_df, SmaxN_small_UI)
}
\arguments{
\item{abund_df}{a numerical dataframe containing the abundance of a 
given species across continuous time for several cameras. The columns refer
to the cameras and the rows refers to the time. \strong{Time must be given
in seconds and be continuous}. \strong{BE CAREFUL that the cameras are 
in the same order in the abund_df and the time_df!}.}

\item{value}{the span of the interval on which the SmaxN should
be computed (given in cell number) ie if in the `time_df` the number is 2
then value is 3 (cells).}

\item{timestep}{the timestep on which the interval begins}

\item{time_df}{a numerical dataframe containing the minimal time 
needed for an individual of the studied species to go from a camera to 
another camera.There are as many rows as there are cameras and there are
as many columns as there are cameras, thus the dataframe is symmetrical
and the diagonal is filled with 0. This dataframe is the output of the 
\code{compute.cam.time} function.}

\item{SmaxN_small_UI}{a numeric value rrefering to the span of the interval
to study.}
}
\value{

}
\description{
This function computes the SmaxN value of the big interval. It choses the 
highest value for each camera taking into account the distance between the
cameras.
}
