% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_Intermediate_functions.R
\name{frame.possible}
\alias{frame.possible}
\title{Compute the frame of possible cells}
\usage{
frame.possible(T, time_df, abund_df, trash_rows)
}
\arguments{
\item{T}{a numerical value referring to the number of the row of the 
abundance dataframe studied (timestep)}

\item{time_df}{a numerical dataframe containing the minimal time 
needed for an individual of the studied species to go from a camera to 
another camera.There are as many rows as there are cameras and there are
as many columns as there are cameras, thus the dataframe is symmetrical
and the diagonal is filled with 0. This dataframe is the output of the 
\code{compute.cam.time} function.}

\item{abund_df}{a numerical dataframe containing the abundance of a 
given species across continuous time for several cameras. The columns refer
to the cameras and the rows refers to the time. \strong{Time must be given
in seconds and be continuous}. \strong{BE CAREFUL that the cameras are 
in the same order in the abund_df and the time_df!}.}

\item{trash_rows}{a vector gathering the number of rows that should not be studied}
}
\value{
the abundance dataframe with NA on all cells except in the frame of 
possible around the studied timestep T.
}
\description{
This function computes for a given timestep, the frame of possible cells:
the cells on which path can be computed. Practically the function adds NA
to the cells not on the frame of possibles in the abund_df.
}
\examples{

abund_df <- data.frame("A" = c(9,8,3,3,3,3,3), "B" = c(0,4,2,2,1,3,3), 
"C" = c(0,0,0,0,1,1,1), "D" = c(1,0,0,0,3,3,3))

time_df <- data.frame("A" = c(0,4,4,3), "B" = c(4,0,2,2), 
                      "C" = c(4,2,0,2), "D" = c(3,2,2,0))
rownames(time_df) <- c("A", "B", "C", "D")

cam_order <- cam.order(time_df)

abund_df <- abund_df[, as.vector(unlist(cam_order))]

frame_possible <- frame.possible(T = 4, time_df, abund_df)

}
