% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_SmaxN_big_interval.R
\name{search.next.value}
\alias{search.next.value}
\title{Give the next value on a given bloc given the first highest values.}
\usage{
search.next.value(max_bloc, bloc, n)
}
\arguments{
\item{max_bloc}{a dataframe containing in rows the number already chosen on
the path and their coordinates ie. on which column (camera name) and row 
(time) they are. Values are ranged fom the biggest (first rows) to the 
lowest (last rows). Column names: value, cam_nm, timestep}

\item{bloc}{a dataframe being the abundance dataframe restricted to the
studied bloc (ie a timestep + span).}

\item{n}{the number of the column on which the max must be computed}
}
\value{
the next value to be chosen that is to say the biggest value of the
bloc after the last value in max_bloc. If there are ex-aequos (ie several
cases of the bloc have the next highest value), the function returns a 
dataframe as max_bloc (Column names: value, cam_nm, timestep) with each row
being an ex-aequo value. If there is no ex-aequo, the function also returns a 
dataframe but with only one row.
}
\description{
This function computes next number value to be chosen in a time bloc. It just gives 
the next highest value, another function informs if this value can be chosen
or not given the time it takes a fish to go between camera pairs.
}
