% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmaxN-package.R
\docType{package}
\name{SmaxN-package}
\alias{SmaxN}
\alias{SmaxN-package}
\title{SmaxN: Compute synchronised maxN when using temporally synchronised cameras}
\description{
The purpose of the `SmaxN` package is to compute an estimation of a species abundance given a network of synchronised cameras.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cmlmagneville.github.io/SmaxN/}
}

}
\author{
\strong{Maintainer}: Camille Magneville \email{camille.magneville@gmail.com} (\href{https://orcid.org/0000-0003-0489-3822}{ORCID})

}
\keyword{internal}
