% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_SmaxN_timestep.R
\name{compute.SmaxN.timestep}
\alias{compute.SmaxN.timestep}
\title{Compute the max of cameras for a given second and a given span}
\usage{
compute.SmaxN.timestep(time_df, abund_df, timestep, value)
}
\arguments{
\item{time_df}{a numerical dataframe containing the minimal time 
needed for an individual of the studied species to go from a camera to 
another camera.There are as many rows as there are cameras and there are
as many columns as there are cameras, thus the dataframe is symmetrical
and the diagonal is filled with 0. This dataframe is the output of the 
\code{compute.cam.time} function}

\item{abund_df}{a numerical dataframe containing the abundance of a 
given species across continuous time for several cameras. The columns refer
to the cameras and the rows refers to the time. \strong{Time must be given
in seconds and be continuous}. \strong{BE CAREFUL that the cameras are 
in the same order in the abund_df and the time_df!}.}

\item{timestep}{the number of the row to the studied timestep}

\item{value}{the value of the span of the interval on which the SmaxN is 
computed. Note: if the span is of 2 seconds, then the interval gathers 
3 cases of the `abund_df`.}
}
\value{
the function returns the SmaxN value for the given timestep and span
}
\description{
This function computes the maximal value of abundance for a given second and 
the "n" seconds after. "n" being set up by the "value" parameter .
}
\examples{
 # Build distance dataframe for the example:
 abund_df_ex <- data.frame("A" = c(0, 1, 3, 7, 2, 2, 3), 
                           "B" = c(2, 2, 2, 2, 0, 0, 0), 
                           "C" = c(2, 0, 1, 0, 0, 4, 2), 
                           "D" = c(0, 1, 0, 1, 0, 6, 1))
 
 # Run the function:
 

}
